DDBTEAM-640: Moved no hits processor into ampq
