+++ src/Message/SearchNoHitsMessage.php
@@ -1,0 +1,14 @@
+<?php
+
+/**
+ * @file
+ */
+
+namespace App\Message;
+
+/**
+ * Class SearchNoHitsMessage.
+ */
+class SearchNoHitsMessage extends BaseMessage
+{
+}
